+++ kwsparentsdk/src/main/java/kws/superawesome/tv/kwsparentsdk/services/oauth/KWSAuthService.java
@@ -2,6 +2,7 @@ package kws.superawesome.tv.kwsparentsdk.services.oauth;
 
 import android.content.Context;
 import android.text.TextUtils;
+import android.util.Log;
 
 import org.json.JSONObject;
 
@@ -11,13 +12,11 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 
 import javax.net.ssl.HttpsURLConnection;
 
-import kws.superawesome.tv.kwsparentsdk.aux.KWSLogger;
 import kws.superawesome.tv.kwsparentsdk.models.oauth.KWSLoggedUser;
 import kws.superawesome.tv.kwsparentsdk.services.KWSHTTPMethod;
 import kws.superawesome.tv.kwsparentsdk.services.KWSService;
@@ -70,20 +69,20 @@ public class KWSAuthService extends KWSService {
         if (success && status == 200 && payload != null) {
             KWSLoggedUser loggedUser = new KWSLoggedUser(payload);
             if (loggedUser.isValid()) {
-                KWSLogger.log("KWSAuthService", "Authenticated OK for user with ID: " + loggedUser.getMetadata().getUserId() + ".");
+                Log.d("KWSAuthService", "Authenticated OK for user with ID: " + loggedUser.getMetadata().getUserId() + ".");
                 listener.didAuthUser(loggedUser);
             } else {
-                KWSLogger.error("KWSAuthService", "User is invalid! Could not auth!");
+                Log.e("KWSAuthService", "User is invalid! Could not auth!");
                 listener.didAuthUser(null);
             }
         } else {
-            KWSLogger.error("KWSAuthService", "Network operation failed trying to Auth user " + email);
+            Log.e("KWSAuthService", "Network operation failed trying to Auth user " + email);
             listener.didAuthUser(null);
         }
 
     }
 
-    public void execute (Context context, String email, String password, KWSInternalAuthInterface listener) {
+    public void execute (Context context, String email, final String password, KWSInternalAuthInterface listener) {
         this.email = email;
         this.password = password;
         this.listener = listener != null ? listener : new KWSInternalAuthInterface() {@Override public void didAuthUser(KWSLoggedUser user) {}};
@@ -92,9 +91,9 @@ public class KWSAuthService extends KWSService {
         final String endpoint = KWSService.kwsApiURL + getEndpoint();
 
         // create a new async task
-        SAAsyncTask task = new SAAsyncTask(context, new SAAsyncTaskInterface() {
+        new SAAsyncTask<>(context, new SAAsyncTaskInterface<SANetworkResult>() {
             @Override
-            public Object taskToExecute() throws Exception {
+            public SANetworkResult taskToExecute() throws Exception {
                 int statusCode;
                 String response;
                 InputStreamReader in;
@@ -160,38 +159,15 @@ public class KWSAuthService extends KWSService {
                 // disconnect
                 conn.disconnect();
 
-                // create the final response
-                HashMap<String, Object> networkResponse = new HashMap<>();
-                networkResponse.put("statusCode", statusCode);
-                networkResponse.put("payload", response);
-
                 // return
-                return networkResponse;
+                return new SANetworkResult(statusCode, response);
             }
 
             @Override
-            public void onFinish(Object result) {
+            public void onFinish(SANetworkResult result) {
 
-                if (result != null && result instanceof HashMap) {
-
-                    // get the hash map
-                    HashMap<String, Object> response = (HashMap<String, Object>)result;
-                    int status = -1;
-                    String payload = null;
-                    if (response.containsKey("statusCode")) {
-                        status = (int) response.get("statusCode");
-                    }
-                    if (response.containsKey("payload")) {
-                        payload = (String) response.get("payload");
-                    }
-
-                    // call the result
-                    if (status > -1 && payload != null) {
-                        success(status, payload, true);
-
-                    } else {
-                        success(status, null, false);
-                    }
+                if (result.isValid()) {
+                    success(result.getStatus(), result.getPayload(), true);
                 } else {
                     success(0, null, false);
                 }
@@ -202,5 +178,58 @@ public class KWSAuthService extends KWSService {
                 success(0, null, false);
             }
         });
+    }
+}
+
+/**
+ * This private class hold the important details needed when receiving a network saDidGetResponse from
+ * a remote server: the HTTP request status (200, 201, 400, 404, etc) and a string saDidGetResponse
+ * that will get parsed subsequently.
+ */
+class SANetworkResult {
+
+    // constants
+    private static final int DEFAULT_STATUS = -1;
+
+    // private variables - status & payload
+    private int status = DEFAULT_STATUS;
+    private String payload;
+
+    /**
+     * Custom constructor taking into account all the class members variables
+     *
+     * @param status    the current request status (200, 201, 400, 404, etc)
+     * @param payload   the current request payload. Can be null
+     */
+    SANetworkResult(int status, String payload) {
+        this.status = status;
+        this.payload = payload;
+    }
+
+    /**
+     * Public getter for the "status" member variable
+     *
+     * @return  the value of the "status" member variable
+     */
+    int getStatus() {
+        return status;
+    }
+
+    /**
+     * Public getter for the "payload" member variable
+     *
+     * @return  the value of the "payload" member variable
+     */
+    String getPayload() {
+        return payload;
+    }
+
+    /**
+     * Method that internally determines if the returned network result has validity
+     *
+     * @return  true or false depending on the condition
+     */
+    boolean isValid () {
+        return status > DEFAULT_STATUS && payload != null;
     }
 }